move utility function to RelatedReplies instead, since it's used by it anyway
